Set serialization to False during testing in settings.py
The SQLite setting in settings.py has been updated for use during testing. With the introduction of a TEST dictionary that sets 'SERIALIZE' to False, we prevent Django from serializing the test database. This allows the test runner to use a test database that is not serialized, improving performance and speeding up automated testing.