:sparkles: feature(Concentration): Add environment-based configurations and deployment improvements
This commit introduces several changes focused on improving development, testing and production environments, as well as streamlining the application deployment process.

Firstly, we added a .env file and integrated python-dotenv into the system for managing environment variables more safely and efficiently.

Secondly, we introduced the 'wait_for_db.sh' script to ensure that the database is ready before the application starts running, adding an additional layer of stability to the system.

Thirdly, we reconfigured the Dockerfile and docker-compose file to support these changes and underline the checks